The current implementation of the Misson Control Store [has a one second ticker](https://github.com/lightningnetwork/lnd/blob/d9887f32129b21e40cf88c79512b16dc5d1715fb/routing/missioncontrol_store.go#L284) that trims the DB, ensuring only the latest 1000 entries are kept.

That ticker was added in https://github.com/lightningnetwork/lnd/pull/5515 as a way to improve performance by batching and de-duplicating the updates to the db.

However, its current implementation can be significantly improved by doing a few pointed changes. In particular:

1. There is no need to open a db tx when there is no work.
2. The ticker does not need to run continuously, only when there are items in the queue that need to be persisted.
3. The current implementation duplicates the entire in-memory cache of the store, which is not ideal when there are only a few entries that need to be added

This PR improves the situation iteratively on each commit. A set of benchmarks is added to verify the performance improvement.

This is the final result (using a bbolt MC store):

```
MissionControlStoreFlushing/no_additional_results-7      221µs ±15%       0µs ± 2%   -99.86%  (p=0.000 n=10+10)
MissionControlStoreFlushing/one_additional_result-7      359µs ±40%      44µs ± 8%   -87.87%  (p=0.000 n=10+10)
MissionControlStoreFlushing/ten_additional_results-7     344µs ± 9%     129µs ±29%   -62.61%  (p=0.000 n=9+10)
MissionControlStoreFlushing/100_additional_results-7     987µs ±28%     621µs ±22%   -37.06%  (p=0.000 n=10+10)
MissionControlStoreFlushing/250_additional_results-7    2.02ms ±37%    1.53ms ±41%   -24.62%  (p=0.005 n=10+10)
MissionControlStoreFlushing/500_additional_results-7    2.88ms ±11%    2.84ms ±18%      ~     (p=0.780 n=9+10)

name                                                  old alloc/op   new alloc/op   delta
MissionControlStoreFlushing/no_additional_results-7      139kB ± 0%       0kB       -100.00%  (p=0.000 n=10+10)
MissionControlStoreFlushing/one_additional_result-7      153kB ± 0%      19kB ± 0%   -87.37%  (p=0.000 n=10+10)
MissionControlStoreFlushing/ten_additional_results-7     175kB ± 0%      41kB ± 0%   -76.35%  (p=0.000 n=8+8)
MissionControlStoreFlushing/100_additional_results-7     350kB ± 0%     222kB ± 0%   -36.63%  (p=0.000 n=10+10)
MissionControlStoreFlushing/250_additional_results-7     651kB ± 0%     535kB ± 0%   -17.75%  (p=0.000 n=8+10)
MissionControlStoreFlushing/500_additional_results-7    1.14MB ± 0%    1.04MB ± 0%    -8.57%  (p=0.000 n=10+10)

name                                                  old allocs/op  new allocs/op  delta
MissionControlStoreFlushing/no_additional_results-7      1.06k ± 0%     0.00k       -100.00%  (p=0.000 n=10+10)
MissionControlStoreFlushing/one_additional_result-7      1.15k ± 0%     0.12k ± 0%   -90.02%  (p=0.000 n=10+10)
MissionControlStoreFlushing/ten_additional_results-7     1.49k ± 0%     0.46k ± 0%   -69.36%  (p=0.000 n=10+10)
MissionControlStoreFlushing/100_additional_results-7     4.90k ± 0%     3.87k ± 0%   -21.07%  (p=0.000 n=10+10)
MissionControlStoreFlushing/250_additional_results-7     10.6k ± 0%      9.6k ± 0%    -9.75%  (p=0.000 n=10+10)
MissionControlStoreFlushing/500_additional_results-7     20.1k ± 0%     19.1k ± 0%    -5.15%  (p=0.000 n=9+10)
```

As can be seen, there is a significant improvement in cpu and memory usage when up to 100 additional results are added (i.e. up to 100 txs/second flowing through the node).

The improvement is specially significant when no new results are added to a full MC store, which is relevant for nodes that have a low or only episodic volument of payments (e.g. a mostly end user node or mobile node).