@matheusd 2024-04-16
@bitromortac sent commit https://github.com/lightningnetwork/lnd/pull/8549/commits/fb4367831fe8766f64e20288a9ab110304ce2f36 to clean up the mission control tests and address your comments.

Also rebased to address the conflict.