@matheusd 2024-05-09
@ellemouton Thank you for the thorough review! I believe I addressed all of your points.

Also, rebased against current master.

Not sure what to do about the release notes, given the v0.18 RCs are already out (move to v0.19 or v0.18.1 or something else).