@ellemouton 2024-05-09
thanks for the quick turn around! I'll check it out soon.

im also not sure re release notes.... cc @saubyk  for that :)