@matheusd 2024-05-16
Removed commit with type safety and moved release notes to v0.19.0.md.