@ellemouton 2024-06-13
@matheusd  - thanks for your patience 🙏  I think we wanna move this to 0.18.1 instead of 0.19 so do you mind doing one last update and moving the release notes to the 0.18.1 doc pls?