@matheusd 2024-06-13
[Updated](https://github.com/lightningnetwork/lnd/compare/e33d572bd222dfe1fd952adbaa8717011802cdd1..8f3dbdc7994ccfe94868d12b9d9168c3e33db526)